En principio ok casos correccion

+++ includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: alejandj <alejandj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/14 12:57:06 by alejandj          #+#    #+#             */
-/*   Updated: 2025/09/28 21:36:06 by alejandj         ###   ########.fr       */
+/*   Updated: 2025/09/29 15:51:11 by alejandj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ int			ft_isdigit(int c);
 long long	get_time_ms(t_sim *sim);
 void		print_status(t_sim *sim, int id, const char *msg);
 void		show_error_args(void);
-void		smart_usleep(t_sim *sim, int time_ms);
+void		smart_usleep(t_sim *sim, long time_ms);
 void		free_forks(t_sim *sim);
 void		clean_up(t_sim *sim);
